+++ src/api/console/login/route.ts
@@ -1,11 +1,21 @@
 import { render } from '#/utils/html.ts'
+import { Layout } from '#/web/components/layout.ts'
 import { Router } from 'express'
-import { Layout } from '#/api/console/components/layout.ts'
 
 const router = Router()
 
 router.get('/', (req, res) => {
-    console.log(req.cookies)
+    if (req.query.code) {
+        return res.redirect(
+            `/auth/authorize?${new URLSearchParams({
+                response_type: 'code',
+                client_id: 'api-panel',
+                code_challenge: 'eIVsW83uLPZmbiKwsR7J86HuUoMqpAWFuoLyo36gpaU', // TODO: dynamic
+                code_challenge_method: 'S256',
+                redirect_uri: 'http://localhost:8080/console/login',
+            })}`,
+        )
+    }
 
     render(
         res,
