+++ src/api/console/mods/route.ts
@@ -1,14 +1,19 @@
 import { html, render } from '#/utils/html.ts'
 import { Router } from 'express'
-import { Layout } from '#/api/console/components/layout.ts'
 import { BUCKETS, s3 } from '#/db/s3.ts'
 import { Upload } from '@aws-sdk/lib-storage'
 import Busboy from 'busboy'
 import type { CompleteMultipartUploadCommandOutput } from '@aws-sdk/client-s3'
+import { getConsoleSession } from '#/utils/api.ts'
+import { Layout } from '#/web/components/layout.ts'
 
 const router = Router()
 
-router.get('/', (_req, res) => {
+router.get('/', async (req, res) => {
+    const session = await getConsoleSession(req)
+    if (!session) {
+        return res.redirect('/console/login')
+    }
     render(
         res,
         Layout({
@@ -18,6 +23,11 @@ router.get('/', (_req, res) => {
 })
 
 router.post('/', async (req, res) => {
+    const session = await getConsoleSession(req)
+    if (!session) {
+        return res.status(401).json({ error: 'Permision Denied' })
+    }
+
     const busboy = Busboy({ headers: req.headers })
 
     let uploadPromise: Promise<CompleteMultipartUploadCommandOutput> | null =
